requirements.txt
# PyYaml==6.0.3
dataclasses==0.6
retry==0.9.2
aiohttp==3.12.15
PyJWT[crypto]==2.10.1
cryptography==46.0.2
orjson==3.11.3